Removing lingering dependencies from dagster-shared to dagster, divest tox suite fully from dagster
Summary:
There were a couple of places where dagster-shared still depended implicitly on dagster (either through missing dependencies or stray code that survived the migration). This PR removes them.

BK

> Insert changelog entry or delete this section.
