+++ python_modules/dagster/dagster/_config/pythonic_config/config.py
@@ -10,6 +10,7 @@ from dagster_shared.dagster_model.pydantic_compat_layer import (
     model_config,
     model_fields,
 )
+from dagster_shared.utils.cached_method import CACHED_METHOD_CACHE_FIELD
 from pydantic import BaseModel, ConfigDict
 from typing_extensions import TypeVar
 
@@ -35,7 +36,6 @@ from dagster._core.errors import (
     DagsterInvalidInvocationError,
     DagsterInvalidPythonicConfigDefinitionError,
 )
-from dagster._utils.cached_method import CACHED_METHOD_CACHE_FIELD
 
 INTERNAL_MARKER = "__internal__"
