+++ python_modules/dagster/dagster_tests/model_tests/test_dagster_model.py
@@ -1,6 +1,6 @@
 import pytest
-from dagster._utils.cached_method import CACHED_METHOD_CACHE_FIELD, cached_method
 from dagster_shared.dagster_model import DagsterModel
+from dagster_shared.utils.cached_method import CACHED_METHOD_CACHE_FIELD, cached_method
 from pydantic import ValidationError
