+++ python_modules/dagster/dagster_tests/utils_tests/test_libraries.py
@@ -1,0 +1,2 @@
+def test_legacy_registry_import_still_works():
+    from dagster._core.libraries import DagsterLibraryRegistry  # noqa
