+++ python_modules/libraries/dagster-shared/dagster_shared/utils/test.py
@@ -1,0 +1,1 @@
+class TestType: ...
