+++ python_modules/libraries/dagster-shared/dagster_shared_tests/test_merger.py
@@ -1,4 +1,4 @@
-from dagster._utils.merger import deep_merge_dicts, merge_dicts
+from dagster_shared.merger import deep_merge_dicts
 
 
 def test_simple_merge():
@@ -55,16 +55,3 @@ def test_realistic():
     }
 
     assert deep_merge_dicts(onto_dict, from_dict) == result_dict
-
-
-def test_merge():
-    # two element merge
-    assert merge_dicts({}, {}) == {}
-    assert merge_dicts({1: 2}, {}) == {1: 2}
-    assert merge_dicts({}, {1: 2}) == {1: 2}
-    assert merge_dicts({1: 1}, {1: 2}) == {1: 2}
-
-    # three element merge
-    assert merge_dicts({}, {}, {}) == {}
-    assert merge_dicts({1: 2}, {2: 3}, {3: 4}) == {1: 2, 2: 3, 3: 4}
-    assert merge_dicts({1: 2}, {1: 3}, {1: 4}) == {1: 4}
