+++ python_modules/libraries/dagster-shared/setup.py
@@ -35,7 +35,12 @@ setup(
     ],
     packages=find_packages(exclude=["dagster_shared_tests*"]),
     include_package_data=True,
-    install_requires=["PyYAML>=5.1"],
+    install_requires=[
+        "PyYAML>=5.1",
+        "packaging>=20.9",
+        "pydantic>=2,<3.0.0",
+        "typing_extensions>=4.10.0,<5",
+    ],
     extras_require={
         "test": ["pytest"],
     },
