+++ python_modules/libraries/dagster-shared/tox.ini
@@ -12,8 +12,6 @@ passenv =
     BUILDKITE*
 install_command = uv pip install {opts} {packages}
 deps =
-  -e ../../dagster[test]
-  -e ../../dagster-pipes
   -e .[test]
 allowlist_externals =
   /bin/bash
